+++ BZI/integration.py
@@ -5,7 +5,7 @@ import numpy as np
 from math import ceil
 from BZI.sampling import HermiteNormalForm
 
-def rectangular_method(PP, grid):
+def rectangular_method(PP, grid, weights):
     """Integrate a pseudopotential within a cell below the Fermi level.
     
     Args:
@@ -20,11 +20,14 @@ def rectangular_method(PP, grid):
     
     integral = 0
     neigvals = int(np.ceil(PP.nvalence_electrons/2.))
-    for kpt in grid:
-        integral += sum(filter(lambda x: x <= PP.fermi_level, PP.eval(kpt, neigvals)))
-    return np.linalg.det(PP.lattice.reciprocal_vectors)/len(grid)*integral
+    for i,kpt in enumerate(grid):
+        integral += weights[i]*sum(filter(lambda x: x <= PP.fermi_level,
+                                          PP.eval(kpt, neigvals)))
+    # return np.linalg.det(PP.lattice.reciprocal_vectors)/len(grid)*integral
+    return np.linalg.det(PP.lattice.reciprocal_vectors)/np.sum(weights)*integral
+    
 
-def rectangular_fermi_level(PP, grid):
+def rectangular_fermi_level(PP, grid, weights):
     """Find the energy at which the toy band structure it cut.
     
     Args:
@@ -36,13 +39,12 @@ def rectangular_fermi_level(PP, grid):
         (float) the Fermi level
     """
     
-    C = int(ceil(PP.nvalence_electrons*len(grid)/2.))
+    C = int(ceil(PP.nvalence_electrons*np.sum(weights)/2.))
     neigvals = int(np.ceil(PP.nvalence_electrons/2)+1)
     energies = np.array([])
-    for g in grid:
-        energies = np.concatenate((energies, PP.eval(g, neigvals)))
+    for i,g in enumerate(grid):
+        energies = np.concatenate((energies, list(PP.eval(g, neigvals))*weights[i]))
     return np.sort(energies)[C-1] # C -1 since python is zero based
-
 
 def monte_carlo(PP, npts):
     """Integrate a function using Monte Carlo sampling. Only works for integrations